Added API for table-rerun
