+++ cli/scripts/ace_exceptions.py
@@ -1,0 +1,2 @@
+class AceException(Exception):
+    pass
